+++ shopping-cart/src/main/java/digitalinnovation/one/experts/shoppingcart/repository/CartRepository.java
@@ -3,8 +3,8 @@ package digitalinnovation.one.experts.shoppingcart.repository;
 import digitalinnovation.one.experts.shoppingcart.model.Cart;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CartRepository extends CrudRepository<Cart, Integer> {
+public interface CartRepository extends CrudRepository<Cart, Long> {
 
-    Cart findByCustomerId(Integer customerId);
+    Cart findByCustomerId(Long customerId);
 
 }
